+++ tensorflow_recommenders/layers/__init__.py
@@ -15,7 +15,8 @@
 # Lint-as: python3
 """Layers."""
 
-from tensorflow_recommenders.layers import dcn
 from tensorflow_recommenders.layers import embedding
 from tensorflow_recommenders.layers import factorized_top_k
+from tensorflow_recommenders.layers import feature_interaction
 from tensorflow_recommenders.layers import loss
+from tensorflow_recommenders.layers.feature_interaction import dcn